move add function to module add
