+++ src/index.js
@@ -59,24 +59,6 @@ import saveIcon from "./images/save1.png"
 // };
 
 
-const add = () => {
-  console.log("add clicked");
-  const input = document.getElementById("inputToDo").value;
-  if (input === ""){return}
-  let toDoList = JSON.parse(localStorage.getItem("allEntries"));
-  let index = null;
-  if (toDoList == null) {
-    index = 1;
-  } else {
-    index = toDoList.length + 1;
-  }
-  const newToDo = new ToDo(input, false, index);
-  const storage = new TodoStorage();
-  storage.addToDo(newToDo);
-  document.getElementById("inputToDo").value = "";
-  toDos();
-};
-
 // const toDos = () => {
 //   clearToDos();
 //   const toDoList = JSON.parse(localStorage.getItem("allEntries"));
@@ -89,24 +71,24 @@ const add = () => {
 // };
 
 
-const edit = (index) => {
-  console.log("edit clicked")
-  toDos();
-  const menuIcon = document.getElementById(`menuBtn${index}`);
-  menuIcon.style.display = "none";
-  const binIcon = document.getElementById(`binBtn${index}`);
-  binIcon.style.display = "block";
-  const saveIcon = document.getElementById(`saveBtn${index}`);
-  saveIcon.style.display = "block"
-  const row = document.getElementById(`toDo${index}`)
-  row.style.background = "rgb(233, 222, 128)"
+// const edit = (index) => {
+//   console.log("edit clicked")
+//   toDos();
+//   const menuIcon = document.getElementById(`menuBtn${index}`);
+//   menuIcon.style.display = "none";
+//   const binIcon = document.getElementById(`binBtn${index}`);
+//   binIcon.style.display = "block";
+//   const saveIcon = document.getElementById(`saveBtn${index}`);
+//   saveIcon.style.display = "block"
+//   const row = document.getElementById(`toDo${index}`)
+//   row.style.background = "rgb(233, 222, 128)"
 
 
-  const p = document.getElementById(`description${index}`)
-  const input = document.getElementById(`descriptionInput${index}`)
-  p.style.display = "none"
-  input.style.display = "block"
-};
+//   const p = document.getElementById(`description${index}`)
+//   const input = document.getElementById(`descriptionInput${index}`)
+//   p.style.display = "none"
+//   input.style.display = "block"
+// };
 window.edit = edit
 
 const save = (index) => {
